Oprava podle mary

+++ src/com/example/stack/StackActiviti.java
@@ -36,12 +36,12 @@ public class StackActiviti extends Activity {
 	public static List<Stack> stackValues;
 
 	@Override
-	protected void onCreate(Bundle savedInstanceState) {
+	public void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.activity_stack_activiti);
 
-		// Button btn = (Button) findViewById(R.id.buttonMenu);
-		// registerForContextMenu(btn);
+		Button btn = (Button) findViewById(R.id.buttonMenu);
+		registerForContextMenu(btn);
 
 		datasource = new StacksDataSource(this);
 		datasource.open();
@@ -57,18 +57,17 @@ public class StackActiviti extends Activity {
 	public void onCreateContextMenu(ContextMenu menu, View v,
 			ContextMenuInfo menuInfo) {
 		super.onCreateContextMenu(menu, v, menuInfo);
-		menu.setHeaderTitle("Context Menu");
-		menu.add(0, v.getId(), 0, "Action 1");
-		menu.add(0, v.getId(), 0, "Action 2");
-
+		menu.setHeaderTitle(getResources().getString(R.string.clearMenu));
+		menu.add(0,1, 0, getResources().getString(R.string.clear));
+		menu.add(0, 100, 0, getResources().getString(R.string.clear_DB));
 	}
 
 	@Override
 	public boolean onContextItemSelected(MenuItem item) {
-		if (item.getTitle() == "Action 1") {
-			Toast.makeText(this, "Action 1 invoked", Toast.LENGTH_SHORT).show();
-		} else if (item.getTitle() == "Action 2") {
-			Toast.makeText(this, "Action 2 invoked", Toast.LENGTH_SHORT).show();
+		if (item.getItemId() == 1) {
+			//clearInput(item.getItemId());
+		} else if (item.getItemId() == 100) {
+		//	clearDB(item.getItemId());
 		} else {
 			return false;
 		}
@@ -367,18 +366,28 @@ public class StackActiviti extends Activity {
 			}
 		}
 
-		public void deleteDatabaseStack() {
-			beTrans();
-			database.execSQL(DB_Helper.DROP_TABLE);
-			database.execSQL(DB_Helper.DATABASE_CREATE);
-			endTrans();
+		public int deleteDatabaseStack() {
+			if (getCountStacks() > 0) {
+				beTrans();
+				database.execSQL(DB_Helper.DROP_TABLE);
+				database.execSQL(DB_Helper.DATABASE_CREATE);
+				endTrans();
+				return 0;
+			} else {
+				return -1;
+			}
 
 		}
 
-		public void clearDatabaseStack() {
-			beTrans();
-			database.delete(DB_Helper.TABLE_STACKS, "1", null);
-			endTrans();
+		public int clearDatabaseStack() {
+			if (getCountStacks() > 0) {
+				beTrans();
+				database.delete(DB_Helper.TABLE_STACKS, "1", null);
+				endTrans();
+				return 0;
+			} else {
+				return -1;
+			}
 
 		}
